Refactor integration tests
This commit creates a new test_utils.go file to group common functions
for tests.
